refactor: migrated to com.zin.jadxaimcp package name
