+++ src/main/resources/META-INF/services/jadx.api.plugins.JadxPlugin
@@ -1,1 +1,1 @@
-jadx.gui.plugins.JadxAIMCP
+com.zin.jadxaimcp.JadxAIMCP
